fix: remove UserConfig
