+++ packages/cli/src/utils/getConfig.ts
@@ -1,6 +1,5 @@
 import type { PossiblePromise } from '@internals/utils'
 import type { CLIOptions, Config } from '@kubb/core'
-import { getCosmiConfig } from './getCosmiConfig.ts'
 
 type ConfigInput = PossiblePromise<Config | Config[]> | ((cli: CLIOptions) => PossiblePromise<Config | Config[]>)
 
@@ -9,18 +8,4 @@ export async function getConfigs(config: ConfigInput, args: CLIOptions): Promise
   const userConfigs = Array.isArray(resolved) ? resolved : [resolved]
 
   return userConfigs.map((item) => ({ ...item, plugins: item.plugins ?? [] }) as Config)
-}
-
-/**
- * Load and return the first Kubb config from the given config file path.
- */
-export async function loadConfig(resolvedConfigPath: string) {
-  const result = await getCosmiConfig(resolvedConfigPath)
-  const configs = await getConfigs(result.config, {})
-
-  if (configs.length === 0) {
-    throw new Error('No configs found')
-  }
-
-  return configs[0]
 }
