+++ packages/core/src/PluginDriver.ts
@@ -330,10 +330,7 @@ export class PluginDriver {
     }
 
     if (gen.operations) {
-      const operationsHandler = async (
-        nodes: Parameters<NonNullable<typeof gen.operations>>[0],
-        ctx: Parameters<NonNullable<typeof gen.operations>>[1],
-      ) => {
+      const operationsHandler = async (nodes: Parameters<NonNullable<typeof gen.operations>>[0], ctx: Parameters<NonNullable<typeof gen.operations>>[1]) => {
         if (ctx.plugin.name !== pluginName) return
         const result = await gen.operations!(nodes, ctx)
         await applyHookResult(result, this, resolveRenderer())
