+++ packages/core/src/index.ts
@@ -26,7 +26,6 @@ export * from './types.ts'
 export type { FunctionParamsAST } from './utils/FunctionParams.ts'
 export { detectFormatter } from './utils/formatters.ts'
 export { getBarrelFiles } from './utils/getBarrelFiles.ts'
-export { getConfigs } from './utils/getConfigs.ts'
 export type { Param, Params } from './utils/getFunctionParams.ts'
 export { createFunctionParams, FunctionParams, getFunctionParams } from './utils/getFunctionParams.ts'
 export { isInputPath } from './utils/isInputPath.ts'
